requirements.txt
Django==4.0.4
django-bootstrap-icons==0.7.8
django-bootstrap-v5==1.0.11
django-icons==21.3
Pillow==9.1.1
psycopg2==2.9.3
djangorestframework==3.13.1